Moved the package json
